% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliff_cal.R
\name{cliff_cal}
\alias{cliff_cal}
\title{Effect size (Cliff's delta) calculation in longdat_disc() pipeline}
\arguments{
\item{melt_data}{Internal function argument.}

\item{Ps_poho_fdr}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{test_var}{Internal function argument.}

\item{data}{Internal function argument.}

\item{verbose}{Internal function argument.}
}
\description{
Effect size (Cliff's delta) calculation in longdat_disc() pipeline
}
