% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConModelTest_cont.R
\name{ConModelTest_cont}
\alias{ConModelTest_cont}
\title{Covariate model test in longdat_cont() pipeline}
\arguments{
\item{N}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{melt_data}{Internal function argument.}

\item{sel_fac}{Internal function argument.}

\item{data_type}{Internal function argument.}

\item{test_var}{Internal function argument.}

\item{verbose}{Internal function argument.}
}
\description{
Covariate model test in longdat_cont() pipeline
}
