% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConModelTest_disc.R
\name{ConModelTest_disc}
\alias{ConModelTest_disc}
\title{Confounding model test in longdat_disc() pipeline}
\arguments{
\item{N}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{melt_data}{Internal function argument.}

\item{sel_fac}{Internal function argument.}

\item{data_type}{Internal function argument.}

\item{test_var}{Internal function argument.}

\item{verbose}{Internal function argument.}
}
\description{
Confounding model test in longdat_disc() pipeline
}
