% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_posthoc.R
\name{correlation_posthoc}
\alias{correlation_posthoc}
\title{Post-hoc test based on correlation test for longdat_cont().}
\usage{
correlation_posthoc(variables, verbose, melt_data, test_var, N)
}
\description{
Post-hoc test based on correlation test for longdat_cont().
}
