% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuneiform_plot.R
\name{cuneiform_plot}
\alias{cuneiform_plot}
\title{Creates cuneiform plots of result table from longdat_disc() or
longdat_cont()}
\usage{
cuneiform_plot(
  result_table,
  x_axis_order = NULL,
  confound_panel = TRUE,
  pos_color = "red",
  neg_color = "blue",
  panel_width = 4,
  title = "LongDat result cuneiform plot",
  title_size = 20,
  confound_text_size = 4,
  x_label_size = 10,
  y_label_size = 10,
  legend_title_size = 12,
  legend_text_size = 10
)
}
\arguments{
\item{result_table}{The result table from longdat_disc() or
longdat_cont() output, or any table that has the same format.}

\item{x_axis_order}{The plotting order of the x axis.
It should be a character vector
(eg. c("Effect_1_2", "Effect_2_3", "Effect_1_3")).}

\item{confound_panel}{A boolean vector indicating whether to plot
confounding status alongside the effect panel. The default is TRUE.}

\item{pos_color}{The color for a positive effect size.
It should be a hex color code (eg. "#b3e6ff") or the colors recognized
by R (see http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf).
The default is "red".}

\item{neg_color}{The color for a negative effect size.
It should be a hex color code (eg. "#b3e6ff") or the colors recognized
by R (see http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf).
The default is "blue".}

\item{panel_width}{The width of the effect size panel on the left
relative to the confounding status panel on the right (width set to 1).
It should be a numerical vector. The default is 4.}

\item{title}{The name of the plot title. The default is
"LongDat result cuneiform plot".}

\item{title_size}{The size of the plot title. The default is 20.}

\item{confound_text_size}{The size of the text in the confounding status
panel. The default is 4.}

\item{x_label_size}{The size of the x label. The default is 10.}

\item{y_label_size}{The size of the y label. The default is 10.}

\item{legend_title_size}{The size of the legend title. The default is 12.}

\item{legend_text_size}{The size of the legend text The default is 10.}
}
\value{
a 'ggplot' object
}
\description{
Creates cuneiform plots of result table from longdat_disc() or
longdat_cont()
}
\details{
This function creates a cuneiform plot which displays the
result of longdat_disc() or longdat_cont(). It plots the effect sizes
 within each time interval for each feature, and also shows the confounding
 status. Only the features with non-NS signals will be included in the plot.
 The output is a ggplot object in patchwork structure. For further
 customization of the plot, please refer to the vignette.
}
\examples{
\dontrun{
# Get the path of example dataset
system.file("Fasting_disc.txt", package = "LongDat")

# Paste the directory to the input below
test_disc <- longdat_disc(input = "your_path_to/Fasting_disc.txt",
data_type = "count", test_var = "Time_point", variable_col = 7,
fac_var = c(1:3))

test_plot <- cuneiform_plot(result_table = test_disc[[1]])
}
}
