% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_p_cal.R
\name{factor_p_cal}
\alias{factor_p_cal}
\title{Calculate the p values for every factor (used for selecting factors later)}
\usage{
factor_p_cal(melt_data, variables, factor_columns, factors, data, N, verbose)
}
\description{
Calculate the p values for every factor (used for selecting factors later)
}
