% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_summarize_cont.R
\name{final_result_summarize_cont}
\alias{final_result_summarize_cont}
\title{Generate result table as output in longdat_cont()}
\usage{
final_result_summarize_cont(
  variable_col,
  N,
  Ps_conf_inv_model_unlist,
  variables,
  sel_fac,
  Ps_conf_model_unlist,
  model_q,
  posthoc_q,
  Ps_null_model_fdr,
  Ps_null_model,
  assoc,
  prevalence,
  mean_abundance,
  p_poho,
  not_used,
  Ps_effectsize,
  data_type,
  false_pos_count
)
}
\arguments{
\item{variable_col}{Internal function argument.}

\item{N}{Internal function argument.}

\item{Ps_conf_inv_model_unlist}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{sel_fac}{Internal function argument.}

\item{Ps_conf_model_unlist}{Internal function argument.}

\item{model_q}{Internal function argument.}

\item{posthoc_q}{Internal function argument.}

\item{Ps_null_model_fdr}{Internal function argument.}

\item{Ps_null_model}{Internal function argument.}

\item{assoc}{Internal function argument.}

\item{prevalence}{Internal function argument.}

\item{mean_abundance}{Internal function argument.}

\item{p_poho}{Internal function argument.}

\item{not_used}{Internal function argument.}

\item{Ps_effectsize}{Internal function argument.}

\item{data_type}{Internal function argument.}

\item{false_pos_count}{Internal function argument.}
}
\description{
Generate result table as output in longdat_cont()
}
