% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_summarize_disc.R
\name{final_result_summarize_disc}
\alias{final_result_summarize_disc}
\title{Generate result table as output in longdat_disc()}
\usage{
final_result_summarize_disc(
  variable_col,
  N,
  Ps_conf_inv_model_unlist,
  variables,
  sel_fac,
  Ps_conf_model_unlist,
  model_q,
  posthoc_q,
  Ps_null_model_fdr,
  Ps_null_model,
  delta,
  case_pairs,
  prevalence,
  mean_abundance,
  Ps_poho_fdr,
  not_used,
  Ps_effectsize,
  case_pairs_name,
  data_type,
  false_pos_count,
  p_wilcox_final
)
}
\arguments{
\item{variable_col}{Internal function argument.}

\item{N}{Internal function argument.}

\item{Ps_conf_inv_model_unlist}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{sel_fac}{Internal function argument.}

\item{Ps_conf_model_unlist}{Internal function argument.}

\item{model_q}{Internal function argument.}

\item{posthoc_q}{Internal function argument.}

\item{Ps_null_model_fdr}{Internal function argument.}

\item{Ps_null_model}{Internal function argument.}

\item{delta}{Internal function argument.}

\item{case_pairs}{Internal function argument.}

\item{prevalence}{Internal function argument.}

\item{mean_abundance}{Internal function argument.}

\item{Ps_poho_fdr}{Internal function argument.}

\item{not_used}{Internal function argument.}

\item{Ps_effectsize}{Internal function argument.}

\item{case_pairs_name}{Internal function argument.}

\item{data_type}{Internal function argument.}

\item{false_pos_count}{Internal function argument.}

\item{p_wilcox_final}{Internal function argument.}
}
\description{
Generate result table as output in longdat_disc()
}
