% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_name_fun.R
\name{fix_name_fun}
\alias{fix_name_fun}
\title{Replace the symbols in variable and confounder names in raw input}
\usage{
fix_name_fun(z)
}
\arguments{
\item{z}{A character vector. This is the character vector
that needs to be changed.}
}
\description{
Replace the symbols in variable and confounder names in raw input
}
