% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongDat_cont_feature_table.R
\docType{data}
\name{LongDat_cont_feature_table}
\alias{LongDat_cont_feature_table}
\title{data/LongDat_cont_feature_table.RData documentation}
\format{
An object of class \code{data.frame} with 20 rows and 4 columns.
}
\usage{
data(LongDat_cont_feature_table)
}
\description{
Example feature data frame for longdat_cont(). This is a dummy data
which contains features (dependent variables).
}
\examples{
\dontrun{
data(LongDat_cont_feature_table)
}
}
\keyword{datasets}
