% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongDat_cont_master_table.R
\docType{data}
\name{LongDat_cont_master_table}
\alias{LongDat_cont_master_table}
\title{data/LongDat_cont_master_table.RData documentation}
\format{
An object of class \code{data.frame} with 20 rows and 9 columns.
}
\usage{
data(LongDat_cont_master_table)
}
\description{
Example master data frame for longdat_cont(). This is a dummy data
which contains metadata and features.
}
\examples{
\dontrun{
data(LongDat_cont_master_table)
}
}
\keyword{datasets}
