% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongDat_disc_master_table.R
\docType{data}
\name{LongDat_disc_master_table}
\alias{LongDat_disc_master_table}
\title{data/LongDat_disc_master_table.RData documentation}
\format{
An object of class \code{data.frame} with 30 rows and 9 columns.
}
\usage{
data(LongDat_disc_master_table)
}
\description{
Example master data frame for longdat_disc(). This is a dummy data
which contains metadata and features.
}
\examples{
\dontrun{
data(LongDat_disc_master_table)
}
}
\keyword{datasets}
