% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongDat_disc_metadata_table.R
\docType{data}
\name{LongDat_disc_metadata_table}
\alias{LongDat_disc_metadata_table}
\title{data/LongDat_disc_metadata_table.RData documentation}
\format{
An object of class \code{data.frame} with 30 rows and 7 columns.
}
\usage{
data(LongDat_disc_metadata_table)
}
\description{
Example metadata data frame for longdat_disc(). This is a dummy data
which contains metadata.
}
\examples{
\dontrun{
data(LongDat_disc_metadata_table)
}
}
\keyword{datasets}
