% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuModelTest_cont.R
\name{NuModelTest_cont}
\alias{NuModelTest_cont}
\title{Null Model Test and post-hoc Test in longdat_cont() pipeline}
\arguments{
\item{N}{Internal function argument.}

\item{data_type}{Internal function argument.}

\item{test_var}{Internal function argument.}

\item{melt_data}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{verbose}{Internal function argument.}
}
\description{
Null Model Test and post-hoc Test in longdat_cont() pipeline
}
