% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuModelTest_disc.R
\name{NuModelTest_disc}
\alias{NuModelTest_disc}
\title{Null Model Test and post-hoc Test in longdat_disc() pipeline}
\usage{
NuModelTest_disc(N, data_type, test_var, melt_data, variables, verbose)
}
\description{
Null Model Test and post-hoc Test in longdat_disc() pipeline
}
