% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue_calc.R
\name{pvalue_calc}
\alias{pvalue_calc}
\title{Calculate the p values from longitudinal data}
\usage{
pvalue_calc(file, p_table_name)
}
\arguments{
\item{file}{This is the data file that has been melted to so that predictors
are created in advance}

\item{p_table_name}{The name you'd like your p_value table to have}
}
\value{
None
}
\description{
Calculate the p values from longitudinal data
}
\examples{
pvalue_calc(file, p_table_name)
}
