% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_sparse_cont.R
\name{rm_sparse_cont}
\alias{rm_sparse_cont}
\title{Remove the dependent variables that are below the threshold of sparsity when the data type is count data in longdat_cont()}
\usage{
rm_sparse_cont(
  values,
  data,
  nonzero_count_cutoff1,
  nonzero_count_cutoff2,
  theta_cutoff,
  Ps_null_model,
  prevalence,
  absolute_sparsity,
  mean_abundance,
  p_poho,
  assoc
)
}
\description{
Remove the dependent variables that are below the threshold of sparsity when the data type is count data in longdat_cont()
}
