% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_calc.R
\name{sig_calc}
\alias{sig_calc}
\title{Determine if the tests are significant based on the previous p value table}
\usage{
sig_calc(file, pair)
}
\arguments{
\item{file}{The p value table file you get from pvalue_calc function}

\item{pair}{Pair is the two time points to compare at. For example, 13 means
timepoint 1 and 3. Here list all the pairs to compare with c(...)}
}
\value{
None
}
\description{
Determine if the tests are significant based on the previous p value table
}
\examples{
sig_calc(file, pair)
}
