% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_table.R
\name{unlist_table}
\alias{unlist_table}
\title{Unlist confound and inverse confound tables, turn them into tables}
\usage{
unlist_table(x, N, variables)
}
\arguments{
\item{x}{The list to be unlisted and turned into table}

\item{N}{Internal function argument.}

\item{variables}{Internal function argument.}
}
\description{
Unlist confound and inverse confound tables, turn them into tables
}
