% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_posthoc.R
\name{wilcox_posthoc}
\alias{wilcox_posthoc}
\title{Wilcoxon post-hoc test}
\usage{
wilcox_posthoc(
  Ps_neg_ctrl_filterd,
  model_q,
  melt_data,
  test_var,
  variables,
  data,
  N,
  verbose
)
}
\description{
Wilcoxon post-hoc test
}
