% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliff_cal.R
\name{cliff_cal}
\alias{cliff_cal}
\title{Effect size (Cliff's delta) calculation in longdat_disc() pipeline}
\usage{
cliff_cal(melt_data, Ps_poho_fdr, variables, test_var, data, verbose)
}
\description{
Effect size (Cliff's delta) calculation in longdat_disc() pipeline
}
