% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConModelTest_cont.R
\name{ConModelTest_cont}
\alias{ConModelTest_cont}
\title{Confounding model test in longdat_cont() pipeline}
\usage{
ConModelTest_cont(
  N,
  variables,
  melt_data,
  sel_fac,
  data_type,
  test_var,
  verbose
)
}
\description{
Confounding model test in longdat_cont() pipeline
}
