% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConModelTest_disc.R
\name{ConModelTest_disc}
\alias{ConModelTest_disc}
\title{Confounding model test in longdat_disc() pipeline}
\usage{
ConModelTest_disc(
  N,
  variables,
  melt_data,
  sel_fac,
  data_type,
  test_var,
  verbose
)
}
\description{
Confounding model test in longdat_disc() pipeline
}
