% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_posthoc.R
\name{correlation_posthoc}
\alias{correlation_posthoc}
\title{Post-hoc test based on correlation test for longdat_cont().}
\usage{
correlation_posthoc(variables, verbose, melt_data, test_var, N)
}
\arguments{
\item{variables}{Internal function argument.}

\item{verbose}{Internal function argument.}

\item{melt_data}{Internal function argument.}

\item{test_var}{Internal function argument.}

\item{N}{Internal function argument.}
}
\description{
Post-hoc test based on correlation test for longdat_cont().
}
