% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_result_summarize_cont.R
\name{final_result_summarize_cont}
\alias{final_result_summarize_cont}
\title{Generate result table as output in longdat_cont()}
\usage{
final_result_summarize_cont(
  variable_col,
  N,
  Ps_conf_inv_model_unlist,
  variables,
  sel_fac,
  Ps_conf_model_unlist,
  model_q,
  posthoc_q,
  Ps_null_model_fdr,
  Ps_null_model,
  assoc,
  case_pairs,
  prevalence,
  mean_abundance,
  p_poho,
  not_used,
  Ps_effectsize,
  output_tag,
  case_pairs_name,
  data_type,
  false_pos_count
)
}
\description{
Generate result table as output in longdat_cont()
}
