% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longdat_cont.R
\name{longdat_cont}
\alias{longdat_cont}
\title{Calculate the p values and effect size from longitudinal data}
\usage{
longdat_cont(input, value_column, factor_column, non_factors)
}
\arguments{
\item{input}{a table with the first column as "Patient", and all the features
should be at the end of the table. Also, place the columns that aren't
numerical (e.g. categorical, ordinal) in the front of the table, right after "Patient"}

\item{value_column}{the number of the position where the features columns start in the table}

\item{factor_column}{the number of the position where the columns that aren't
numerical  (e.g. categorical, ordinal) ends in the table}

\item{non_factors}{the factors that are not considered as potential confounders to the factor tested for,
for example, "Patient" and "SampleID" are random factors instead of confounders, so
they should be listed as non_factors}
}
\description{
Calculate the p values and effect size from longitudinal data
}
\examples{
longdat_disc(table, 24, 5, c("Case", "Patient", "SampleID"))
}
