% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NuModelTest_cont.R
\name{NuModelTest_cont}
\alias{NuModelTest_cont}
\title{Null Model Test and post-hoc Test in longdat_cont() pipeline}
\usage{
NuModelTest_cont(N, data_type, test_var, melt_data, variables, verbose)
}
\description{
Null Model Test and post-hoc Test in longdat_cont() pipeline
}
