% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_sparse_cont.R
\name{rm_sparse_cont}
\alias{rm_sparse_cont}
\title{Remove the dependent variables that are below the threshold of
sparsity when the data type is count data in longdat_cont()}
\arguments{
\item{values}{Internal function argument.}

\item{data}{Internal function argument.}

\item{nonzero_count_cutoff1}{Internal function argument.}

\item{nonzero_count_cutoff2}{Internal function argument.}

\item{theta_cutoff}{Internal function argument.}

\item{Ps_null_model}{Internal function argument.}

\item{prevalence}{Internal function argument.}

\item{absolute_sparsity}{Internal function argument.}

\item{mean_abundance}{Internal function argument.}

\item{p_poho}{Internal function argument.}

\item{assoc}{Internal function argument.}
}
\description{
Remove the dependent variables that are below the threshold of
sparsity when the data type is count data in longdat_cont()
}
