% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_posthoc.R
\name{wilcox_posthoc}
\alias{wilcox_posthoc}
\title{Wilcoxon post-hoc test}
\usage{
wilcox_posthoc(
  result_neg_ctrl,
  model_q,
  melt_data,
  test_var,
  variables,
  data,
  N,
  verbose
)
}
\arguments{
\item{result_neg_ctrl}{Internal function argument.}

\item{model_q}{Internal function argument.}

\item{melt_data}{Internal function argument.}

\item{test_var}{Internal function argument.}

\item{variables}{Internal function argument.}

\item{data}{Internal function argument.}

\item{N}{Internal function argument.}

\item{verbose}{Internal function argument.}
}
\description{
Wilcoxon post-hoc test
}
